autocall can now work with service maps.
