+++ include/kangaru/container.hpp
@@ -1,0 +1,228 @@
+#pragma once
+
+#include "detail/function_traits.hpp"
+#include "detail/utils.hpp"
+#include "detail/container_service.hpp"
+
+#include <unordered_map>
+#include <memory>
+#include <type_traits>
+#include <tuple>
+#include <vector>
+#include <exception>
+
+namespace kgr {
+
+struct Container {
+private:
+	template<typename Condition, typename T = detail::enabler> using enable_if = detail::enable_if_t<Condition::value, T>;
+	template<typename Condition, typename T = detail::enabler> using disable_if = detail::enable_if_t<!Condition::value, T>;
+	template<typename T> using decay = typename std::decay<T>::type;
+	template<typename T> using is_single = std::is_base_of<Single, decay<T>>;
+	template<typename T> using is_base_of_container_service = std::is_base_of<detail::ContainerServiceBase, T>;
+	template<typename T> using is_base_of_container = std::is_base_of<Container, T>;
+	template<typename T> using is_container_service = std::is_same<T, ContainerService>;
+	template<typename T> using parent_types = typename decay<T>::ParentTypes;
+	template<typename T> using is_abstract = std::is_abstract<T>;
+	template<int S, typename T> using parent_element = typename std::tuple_element<S, parent_types<T>>::type;
+	template<typename Tuple> using tuple_seq = typename detail::seq_gen<std::tuple_size<Tuple>::value>::type;
+	template<typename Tuple, int n> using tuple_seq_minus = typename detail::seq_gen<std::tuple_size<Tuple>::value - n>::type;
+	template<typename T> using instance_ptr = std::unique_ptr<T, void(*)(void*)>;
+	
+	constexpr static detail::enabler null = detail::null;
+	
+	template<typename T>
+	static void deleter(void* i) {
+		delete static_cast<T*>(i);
+	}
+	
+public:
+	Container(const Container &) = delete;
+	Container& operator =(const Container &) = delete;
+	
+	Container(Container&& other) : _instances{std::move(other._instances)}, _services{std::move(other._services)} {}
+	
+	Container& operator =(Container&& other) {
+		std::swap(other._instances, _instances);
+		std::swap(other._services, _services);
+		
+		return *this;
+	}
+	
+	virtual ~Container() = default;
+
+	template <typename T, typename... Args, enable_if<is_base_of_container<T>> = null>
+	static T make_container(Args&&... args) {
+		T container{std::forward<Args>(args)...};
+		container.init();
+		
+		return container;
+	}
+
+	template<typename T>
+	void instance(T&& service) {
+		static_assert(is_single<decay<T>>::value, "instance() only accept Single Service instance.");
+		
+		save_instance(std::forward<T>(service));
+	}
+	
+	template<typename T, typename... Args>
+	void instance(Args&& ...args) {
+		static_assert(is_single<T>::value, "instance() only accept Single Service instance.");
+		save_instance(make_service_instance<T>(std::forward<Args>(args)...));
+	}
+	
+	template<typename T, typename... Args, disable_if<is_base_of_container<T>> = null>
+	ServiceType<T> service(Args&& ...args) {
+		return get_service<T>(std::forward<Args>(args)...).forward();
+	}
+	
+	template<typename U, typename ...Args>
+	detail::function_result_t<decay<U>> invoke(U&& function, Args&&... args) {
+		return invoke_helper(tuple_seq_minus<detail::function_arguments_t<decay<U>>, sizeof...(Args)>{}, std::forward<U>(function), std::forward<Args>(args)...);
+	}
+	
+	template<template<typename> class Map, typename U, typename ...Args>
+	detail::function_result_t<decay<U>> invoke(U&& function, Args&&... args) {
+		return invoke_helper<Map>(tuple_seq_minus<detail::function_arguments_t<decay<U>>, sizeof...(Args)>{}, std::forward<U>(function), std::forward<Args>(args)...);
+	}
+	
+protected:
+	Container() = default;
+	void init(){}
+	
+private:
+	// save instance functions
+	template<typename T, enable_if<detail::has_overrides<decay<T>>> = null>
+	void save_instance(T&& service) {
+		save_instance(std::forward<T>(service), tuple_seq<parent_types<T>>{});
+	}
+	
+	template<typename T, disable_if<detail::has_overrides<decay<T>>> = null>
+	void save_instance(T&& service) {
+		using U = decay<T>;
+		save_instance_helper<U>(instance_ptr<U>{new U{std::move(service)}, &Container::deleter<U>});
+	}
+	
+	template<typename T, int... S>
+	void save_instance(T&& service, detail::seq<S...>) {
+		using U = decay<T>;
+		save_instance_helper<U, parent_element<S, U>...>(instance_ptr<U>{new U{std::move(service)}, &Container::deleter<U>});
+	}
+	
+	template<typename T>
+	void save_instance_helper(instance_ptr<T> service) {
+		_services[detail::type_id<T>] = service.get();
+		_instances.emplace_back(std::move(service));
+	}
+	
+	template<typename T, typename Override, typename... Others>
+	void save_instance_helper(instance_ptr<T> service) {
+		using ServiceOverride = detail::ServiceOverride<T, Override>;
+
+		auto baseService = instance_ptr<Override>{new ServiceOverride{*service.get()}, &Container::deleter<ServiceOverride>};
+		_services[detail::type_id<Override>] = baseService.get();
+		_instances.emplace_back(std::move(baseService));
+		save_instance_helper<T, Others...>(std::move(service));
+	}
+	
+	// get service functions
+	template<typename T, typename... Args, disable_if<is_single<T>> = null, disable_if<is_base_of_container<T>> = null>
+	T get_service(Args ...args) {
+		auto service = make_service_instance<T>(std::forward<Args>(args)...);
+		invoke_service(service);
+		return std::move(service);
+	}
+	
+	template<typename T, enable_if<is_container_service<T>> = null>
+	T get_service() {
+		return T{*this};
+	}
+	
+	template<typename T, disable_if<is_container_service<T>> = null, enable_if<is_base_of_container_service<T>> = null>
+	T get_service() {
+		return T{*dynamic_cast<typename T::Type*>(this)};
+	}
+	
+	template<typename T, enable_if<is_abstract<T>> = null, enable_if<is_single<T>> = null, disable_if<is_base_of_container<T>> = null>
+	T& get_service() {
+		auto it = _services.find(detail::type_id<T>);
+		
+		if (it != _services.end()) {
+			return *static_cast<T*>(it->second);
+		}
+		
+		throw std::runtime_error("No service instance can be used as abstract service");
+	}
+	
+	template<typename T, enable_if<is_single<T>> = null, disable_if<is_base_of_container<T>> = null, disable_if<is_abstract<T>> = null>
+	T& get_service() {
+		auto it = _services.find(detail::type_id<T>);
+		
+		if (it == _services.end()) {
+			instance(make_service_instance<T>());
+			
+			auto service = static_cast<T*>(_services[detail::type_id<T>]);
+			invoke_service(*service);
+			return *service;
+		} else {
+			return *static_cast<T*>(it->second);
+		}
+	}
+	
+	// make instance
+	template<typename T, typename... Args>
+	T make_service_instance(Args&&... args) {
+		return invoke(&T::construct, std::forward<Args>(args)...);
+	}
+	
+	// invoke
+	template<typename U, typename ...Args, int... S>
+	detail::function_result_t<decay<U>> invoke_helper(detail::seq<S...>, U&& function, Args&&... args) {
+		return function(get_service<decay<detail::function_argument_t<S, decay<U>>>>()..., std::forward<Args>(args)...);
+	}
+	
+	template<template<typename> class Map, typename U, typename ...Args, int... S>
+	detail::function_result_t<decay<U>> invoke_helper(detail::seq<S...>, U&& function, Args&&... args) {
+		return function(service<typename Map<detail::function_argument_t<S, decay<U>>>::Service>()..., std::forward<Args>(args)...);
+	}
+	
+	template<typename T, int... S, enable_if<detail::has_invoke<decay<T>>> = null>
+	void invoke_service(T&& service) {
+		using U = decay<T>;
+		invoke_service_helper<typename U::invoke::Next>(std::forward<T>(service), U::invoke::method);
+	}
+	
+	template<typename Method, typename T, typename F, enable_if<detail::has_next<Method>> = null>
+	void invoke_service_helper(T&& service, F&& function) {
+		do_invoke_service(tuple_seq<detail::function_arguments_t<decay<F>>>{}, std::forward<T>(service), std::forward<F>(function));
+		invoke_service_helper<typename Method::Next>(std::forward<T>(service), Method::method);
+	}
+	
+	template<typename Method, typename T, typename F, disable_if<detail::has_next<Method>> = null>
+	void invoke_service_helper(T&& service, F&& function) {
+		do_invoke_service(tuple_seq<detail::function_arguments_t<decay<F>>>{}, std::forward<T>(service), std::forward<F>(function));
+	}
+	
+	template<typename T, typename F, int... S>
+	void do_invoke_service(detail::seq<S...>, T&& service, F&& function) {
+		invoke([&service, &function](detail::function_argument_t<S, decay<F>>... args){
+			(service.*function)(std::forward<detail::function_argument_t<S, decay<F>>>(args)...);
+		});
+	}
+	
+	template<typename T, disable_if<detail::has_invoke<decay<T>>> = null>
+	void invoke_service(T&&) {}
+	
+	std::vector<instance_ptr<void>> _instances;
+	std::unordered_map<detail::type_id_t, void*> _services;
+};
+
+template<typename T = Container, typename ...Args>
+T make_container(Args&& ...args) {
+	static_assert(std::is_base_of<Container, T>::value, "make_container only accept container types.");
+
+	return Container::make_container<T>(std::forward<Args>(args)...);
+}
+
+}  // namespace kgr
